+++ src/components/dashboard/dashboard.component.ts
@@ -1,20 +1,21 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnDestroy, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 
 import { CreateComponent } from '../../elements/create/create.component';
 import { DialogComponent } from '../../elements/dialog/dialog.component';
 import { Superhero } from '../../interfaces/superhero.interface';
 import { DataService } from '../../services/data.service'; 
-import { Observable, of } from 'rxjs';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.scss']
 })
-export class DashboardComponent implements OnInit {
+export class DashboardComponent implements OnDestroy, OnInit {
   public columnsToDisplay: string[];
-  public superheroes$: Observable<Superhero[]>;
+  public superheroes: Superhero[];
+  public superheroesSubscription: Subscription;
 
   constructor(
     public readonly dialog: MatDialog,
@@ -26,22 +27,27 @@ export class DashboardComponent implements OnInit {
     this.columnsToDisplay = ['nameLabel', 'genderLabel', 'citizenshipLabel', 'skillsLabel', 'occupationLabel', 'memberOfLabel', 'creatorLabel'];
 
     // Get superheroes from the service
-    this.superheroes$ = this.dataService.getSuperheroes();
+    this.superheroesSubscription = this.dataService.getSuperheroes().subscribe( (heroesList: Superhero[]) => {
+      this.superheroes = heroesList;
+    });
   }
 
-  public updateList(newList$: Observable<Superhero[]>): void {
-    this.superheroes$ = newList$;
+  ngOnDestroy(): void {
+    this.superheroesSubscription?.unsubscribe();
   }
 
-  public openDialog(superheroSelected: Superhero): void {
-    const dialogRef = this.dialog.open(DialogComponent, {data: superheroSelected });
+  public updateList(newList: Superhero[]): void {
+    this.superheroes = newList;
+    this.forceTableReload();
+  }
+
+  public openInfoDialog(superheroSelected: Superhero): void {
+    const dialogRef = this.dialog.open(DialogComponent, { data: superheroSelected });
 
     dialogRef.afterClosed().subscribe( (result: string) => {
       switch (result) {
         case 'delete':
-          this.superheroes$.subscribe((heroes: Superhero[]) => {
-            this.superheroes$ = of(heroes.filter((hero: Superhero) => hero.nameLabel !== superheroSelected.nameLabel));
-          });
+          this.superheroes = this.superheroes.filter((hero: Superhero) => hero.nameLabel !== superheroSelected.nameLabel);
           break;
         case 'edit':
           this.edit(superheroSelected);
@@ -58,9 +64,8 @@ export class DashboardComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe( (result: Superhero) => {
       if (result) {
-        this.superheroes$.subscribe((heroes: Superhero[]) => {
-          this.superheroes$ = of([result, ...heroes]);
-        });
+        this.superheroes = [result, ...this.superheroes];
+        this.forceTableReload();
       }
     });
   }
@@ -74,12 +79,15 @@ export class DashboardComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe( (result: Superhero) => {
       if (result) {
-        this.superheroes$.subscribe((heroes: Superhero[]) => {
-          const index = heroes.findIndex((hero: Superhero) => hero.nameLabel === result.nameLabel);
-          heroes[index] = result;
-          this.superheroes$ = of(heroes);
-        });
+        const index = this.superheroes.findIndex((hero: Superhero) => hero.nameLabel === result.nameLabel);
+        this.superheroes[index] = result;
+        this.forceTableReload();
       }
     });
+  }
+
+  // Trick to force ngOnChanges into table component and reload the table content
+  private forceTableReload(): void {
+    this.superheroes = [...this.superheroes];
   }
 }