+++ src/elements/chips/chips.component.ts
@@ -1,8 +1,8 @@
 import { COMMA, ENTER } from '@angular/cdk/keycodes';
-import { Component, ElementRef, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
+import { Component, ElementRef, EventEmitter, Input, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { MatAutocompleteSelectedEvent } from '@angular/material/autocomplete';
-import { Observable, of } from 'rxjs';
+import { Observable } from 'rxjs';
 import { map, startWith } from 'rxjs/operators';
 
 import { Superhero } from '../../interfaces/superhero.interface';
@@ -17,34 +17,32 @@ export class ChipsComponent implements OnInit {
   public formControl = new FormControl();
   public filteredSuperheroes$: Observable<Superhero[]>;
   public superheroesList: Superhero[] = [];
-  public allSuperheroes: Superhero[];
+  public allSuperheroes: Superhero[] = [];
 
   @ViewChild('inputRef', {static: true}) inputRef: ElementRef<HTMLInputElement>;
 
-  @Input() superheroes$: Observable<Superhero[]>;
-  @Output() updateList = new EventEmitter<Observable<Superhero[]>>();
+  @Input() superheroes: Superhero[];
+  @Output() updateList = new EventEmitter<Superhero[]>();
 
   ngOnInit(): void {
-    this.superheroes$.subscribe( (data: Superhero[]) => {
-      this.allSuperheroes = data;
+    this.allSuperheroes = [...this.superheroes];
 
-      this.filteredSuperheroes$ = this.formControl.valueChanges.pipe(
-        startWith(null),
-        map((superheroe: string | Superhero | null) =>
-          superheroe ? this._filterByName(superheroe) : this.allSuperheroes.slice()
-        )
-      );
-    })
+    this.filteredSuperheroes$ = this.formControl.valueChanges.pipe(
+      startWith(null),
+      map((superheroe: string | Superhero | null) =>
+        superheroe ? this._filterByName(superheroe) : this.allSuperheroes.slice()
+      )
+    );
   }
 
   public remove(superhero: Superhero): void {
     this.superheroesList = this.superheroesList.filter(((value: Superhero) => value !== superhero));
 
-    // If all the filters have been cleared returnt the full list
+    // If all the filters have been cleared return the full list
     if (this.superheroesList.length === 0) {
-      this.updateList.emit(of(this.allSuperheroes));
+      this.updateList.emit(this.allSuperheroes);
     } else {
-      this.updateList.emit(of(this.superheroesList));
+      this.updateList.emit(this.superheroesList);
     }
   }
 
@@ -52,7 +50,7 @@ export class ChipsComponent implements OnInit {
     const selectedSuperheroe = this.allSuperheroes.find((superhero) => superhero.nameLabel === event.option.value.nameLabel);
     if (selectedSuperheroe) {
       this.superheroesList.push(selectedSuperheroe);
-      this.updateList.emit(of(this.superheroesList));
+      this.updateList.emit(this.superheroesList);
     }
 
     // Clear input (both lines are needed because a known issue in the component)